requirements/base.txt
Pillow==10.2.0
redis==5.0.3
celery==5.3.6
flower==2.0.1
stripe==4.0.2
WeasyPrint==52.5

# Django
# ------------------------------------------------------------------------------
Django==4.2.10
django-environ==0.11.2
